@@ -14,17 +14,17 @@
   },
   "devDependencies": {
     "@eslint-react/eslint-plugin": "workspace:*",
-    "@eslint/js": "9.0.0",
+    "@eslint/js": "9.1.1",
     "@tsconfig/node20": "20.1.4",
     "@tsconfig/strictest": "2.0.5",
-    "@types/react": "18.2.78",
+    "@types/react": "18.2.79",
     "@types/react-dom": "18.2.25",
     "@vitejs/plugin-react": "4.2.1",
-    "eslint": "9.0.0",
+    "eslint": "9.1.0",
     "eslint-plugin-react-hooks": "4.6.0",
     "eslint-plugin-react-refresh": "0.4.6",
     "typescript": "5.4.5",
-    "typescript-eslint": "7.6.0",
-    "vite": "5.2.8"
+    "typescript-eslint": "7.7.0",
+    "vite": "5.2.10"
   }
 }
