@@ -1,6 +1,6 @@
 {
   "name": "@eslint-react/monorepo",
-  "version": "1.5.9-beta.0",
+  "version": "1.5.10-beta.0",
   "private": true,
   "description": "A series of composable ESLint rules for libraries and frameworks that use React as a UI runtime.",
   "keywords": [
@@ -34,7 +34,7 @@
     "build:all": "pnpm run build && pnpm run build:docs && pnpm run build:website",
     "build:docs": "turbo build:docs --filter='./packages/**'",
     "build:force": "turbo build --force --filter='./packages/**'",
-    "build:website": "cd website && pnpm run build",
+    "build:website": "pnpm -F './website' run build",
     "format:check": "dprint check",
     "format:package-json": "bun run --bun ./scripts/format-package-json.ts",
     "format:write": "dprint fmt",
@@ -47,7 +47,7 @@
     "lint:spell": "cspell lint --relative --no-progress '**'",
     "lint:ts": "eslint --cache .",
     "lint:type": "pnpm -r run --parallel lint:type",
-    "lint:website": "cd website && pnpm run lint",
+    "lint:website": "pnpm -F './website' run lint",
     "prepare": "husky install && pnpm run build",
     "test": "vitest --run --logHeapUsage",
     "test:ui": "vitest --ui",
@@ -59,27 +59,27 @@
     "@tsconfig/node20": "20.1.4",
     "@tsconfig/strictest": "2.0.5",
     "@types/node": "20.12.7",
-    "@typescript-eslint/eslint-plugin": "7.6.0",
-    "@typescript-eslint/parser": "7.6.0",
-    "@typescript-eslint/rule-tester": "^7.6.0",
+    "@typescript-eslint/eslint-plugin": "7.7.0",
+    "@typescript-eslint/parser": "7.7.0",
+    "@typescript-eslint/rule-tester": "^7.7.0",
     "@vitest/ui": "1.5.0",
-    "bun": "1.1.3",
-    "bun-types": "1.1.3",
+    "bun": "1.1.4",
+    "bun-types": "1.1.4",
     "code-block-writer": "13.0.1",
     "cspell": "8.7.0",
     "dedent": "1.5.3",
     "dprint": "0.45.1",
-    "effect": "2.4.18",
-    "eslint": "9.0.0",
+    "effect": "3.0.3",
+    "eslint": "9.1.0",
     "eslint-config-flat-gitignore": "0.1.5",
     "eslint-plugin-eslint-comments": "3.2.0",
     "eslint-plugin-eslint-plugin": "6.0.0",
     "eslint-plugin-filenames-simple": "0.9.0",
-    "eslint-plugin-functional": "6.4.1",
+    "eslint-plugin-functional": "6.5.1",
     "eslint-plugin-jsdoc": "48.2.3",
     "eslint-plugin-perfectionist": "2.9.0",
     "eslint-plugin-simple-import-sort": "12.1.0",
-    "eslint-plugin-vitest": "0.5.2",
+    "eslint-plugin-vitest": "0.5.3",
     "fast-equals": "5.0.1",
     "husky": "9.0.11",
     "markdownlint": "0.34.0",
@@ -87,7 +87,7 @@
     "publint": "0.2.7",
     "skott": "0.33.2",
     "sort-package-json": "2.10.0",
-    "taze": "0.13.4",
+    "taze": "0.13.6",
     "tiny-invariant": "1.3.3",
     "ts-pattern": "5.1.1",
     "tsup": "8.0.2",
@@ -99,7 +99,7 @@
     "typescript": "5.4.5",
     "vitest": "1.5.0"
   },
-  "packageManager": "pnpm@8.15.4",
+  "packageManager": "pnpm@9.0.4",
   "engines": {
     "bun": ">=1.0.15",
     "node": ">=18.18.0"
@@ -114,7 +114,7 @@
       "array.prototype.toreversed": "npm:@nolyfill/array.prototype.toreversed@latest",
       "array.prototype.tosorted": "npm:@nolyfill/array.prototype.tosorted@latest",
       "es-iterator-helpers": "npm:@nolyfill/es-iterator-helpers@latest",
-      "eslint": "9.0.0",
+      "eslint": ">=9.1.0",
       "function-bind": "npm:@nolyfill/function-bind@latest",
       "has": "npm:@nolyfill/has@latest",
       "hasown": "npm:@nolyfill/hasown@latest",
