@@ -9,30 +9,30 @@
     "start": "next start"
   },
   "dependencies": {
-    "next": "14.2.1",
-    "nextra": "3.0.0-alpha.22",
-    "nextra-theme-docs": "3.0.0-alpha.22",
+    "next": "14.2.2",
+    "nextra": "3.0.0-alpha.23",
+    "nextra-theme-docs": "3.0.0-alpha.23",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "ts-pattern": "5.1.1"
   },
   "devDependencies": {
     "@eslint-react/eslint-plugin": "workspace:*",
-    "@eslint/js": "9.0.0",
+    "@eslint/js": "9.1.1",
     "@mdx-js/mdx": "3.0.1",
     "@tsconfig/node20": "20.1.4",
     "@tsconfig/strictest": "2.0.5",
     "@types/node": "20.12.7",
-    "@types/react": "18.2.78",
+    "@types/react": "18.2.79",
     "@types/react-dom": "18.2.25",
-    "@typescript-eslint/eslint-plugin": "7.6.0",
-    "@typescript-eslint/parser": "7.6.0",
+    "@typescript-eslint/eslint-plugin": "7.7.0",
+    "@typescript-eslint/parser": "7.7.0",
     "@vanilla-extract/css": "1.14.2",
     "@vanilla-extract/next-plugin": "2.4.0",
     "@vitejs/plugin-react": "4.2.1",
     "autoprefixer": "10.4.19",
-    "effect": "2.4.18",
-    "eslint": "9.0.0",
+    "effect": "3.0.3",
+    "eslint": "9.1.0",
     "eslint-plugin-mdx": "3.1.5",
     "eslint-plugin-react-hooks": "4.6.0",
     "eslint-plugin-react-refresh": "0.4.6",
@@ -43,7 +43,7 @@
     "remark-gfm": "4.0.0",
     "sass": "1.75.0",
     "typescript": "5.4.5",
-    "typescript-eslint": "7.6.0",
-    "vite": "5.2.8"
+    "typescript-eslint": "7.7.0",
+    "vite": "5.2.10"
   }
 }
