@@ -16,7 +16,7 @@ export default createRule<[], MessageID>({
       requiresTypeChecking: false,
     },
     messages: {
-      ENSURE_CUSTOM_HOOKS_USING_OTHER_HOOKS: "Custom hooks {{name}} should use other hooks",
+      ENSURE_CUSTOM_HOOKS_USING_OTHER_HOOKS: "A custom hook {{name}} should use other hooks",
     },
     schema: [],
   },
