@@ -66,8 +66,8 @@ export default createRule<Options, MessageID>({
       requiresTypeChecking: false,
     },
     messages: {
-      FILE_NAME_EXTENSION_INVALID: "JSX file extension is required",
-      FILE_NAME_EXTENSION_UNEXPECTED: "use JSX file extension as needed",
+      FILE_NAME_EXTENSION_INVALID: "The JSX file extension is required",
+      FILE_NAME_EXTENSION_UNEXPECTED: "Use JSX file extension as needed",
     },
     schema,
   },
