@@ -75,10 +75,10 @@ export default createRule<Options, MessageID>({
       requiresTypeChecking: false,
     },
     messages: {
-      FILENAME_CASE_MISMATCH: "File name '{{name}}' does not match {{rule}}",
+      FILENAME_CASE_MISMATCH: "A file with name '{{name}}' does not match {{rule}}",
       FILENAME_CASE_MISMATCH_SUGGESTION:
-        "File name '{{name}}' does not match {{rule}}. Should rename to '{{suggestion}}'.",
-      FILENAME_EMPTY: "File name is empty",
+        "A file with name '{{name}}' does not match {{rule}}. Should rename to '{{suggestion}}'.",
+      FILENAME_EMPTY: "A file must have non-empty name",
     },
     schema,
   },
