@@ -1,7 +1,7 @@
+import tsEsLintParser from "@typescript-eslint/parser";
 import { RuleTester } from "@typescript-eslint/utils/ts-eslint";
 import path from "pathe";
 import * as vitest from "vitest";
-import tsEsLintParser from "@typescript-eslint/parser";
 
 RuleTester.it = vitest.it;
 RuleTester.itOnly = vitest.it.only;
@@ -12,14 +12,14 @@ export function getFixturesRootDir(): string {
 }
 
 export const defaultLanguageOptions = {
-  sourceType: "module",
   ecmaVersion: "latest",
   parser: tsEsLintParser,
   parserOptions: {
-    tsconfigRootDir: getFixturesRootDir(),
-    project: true,
     ecmaFeatures: { jsx: true },
+    project: true,
+    tsconfigRootDir: getFixturesRootDir(),
   },
+  sourceType: "module",
 } as const;
 
 export const ruleTester = new RuleTester({ languageOptions: defaultLanguageOptions } as never);
